backend_python-v2/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
filelock==3.16.1
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==3.0.2
mysqlclient==2.2.6
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==1.7.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
